% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{adj_R2}
\alias{adj_R2}
\title{adjR2: Calculate the adjusted r-squared statistic between predicted and ground truth continuous variables.}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{A numeric vector of predicted values.}

\item{truth}{A numeric vector of ground truth values.}

\item{n_p}{The number of model parameters, excluding the intercept.}
}
\value{
The adjusted R-squared statistic between the predicted and ground truth values.
}
\description{
This function calculates the adjusted r-squared statistic between the predicted values and
the ground truth continuous variables, taking into account the number of model parameters.
}
\examples{
pred = c(2, 4, 6, 8)
truth = c(2.2, 3.8, 6.1, 7.9)
n_p = 2
adjR2(pred, truth, n_p)
}
