% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{classes}
\alias{classes}
\title{classes: Get classes of each variable in a tibble.}
\usage{
classes(x)
}
\arguments{
\item{x}{A tibble or data frame.}
}
\value{
A character vector containing the classes of each variable in the input tibble or data frame.
}
\description{
This function returns a character vector containing the classes of each variable in a tibble \code{x}.
It is similar to the \code{names} function but returns the classes instead of the variable names.
}
\examples{
library(tibble)
df = tibble(A = 1:3, B = c("a", "b", "c"), C = factor(c("X", "Y", "Z")))
classes(df)
}
