% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{df_scale}
\alias{df_scale}
\title{dF_scale: Scale numeric variables in a tibble.}
\usage{
df_scale(x, center = T, scale = T)
}
\arguments{
\item{x}{A tibble or data frame.}

\item{center}{Logical; should the numeric variables be centered (default is TRUE).}

\item{scale}{Logical; should the numeric variables be scaled (default is TRUE).}
}
\value{
A tibble with the numeric variables centered and/or scaled.
}
\description{
This function returns a tibble \code{x} in which the numeric variables have been
centered and/or scaled using the \code{scale} function. It retains the variable
attributes for more informative output.
}
\examples{
library(tibble)
df = tibble(A = 1:3, B = c("a", "b", "c"), C = factor(c("X", "Y", "Z")), D = c(10, 20, 30))
df_scale(df)
}
