% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_max}
\alias{is_max}
\title{is_max: Check if elements of an atomic vector are equal to its maximum value.}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector.}

\item{na.rm}{Logical; should NA values be removed before calculating the maximum value (default is TRUE).}
}
\value{
A logical vector with TRUE where \code{x} equals its maximum value.
}
\description{
This function accepts an atomic vector \code{x} and returns a logical vector with
TRUE where \code{x} equals its maximum value.
}
\examples{
vec = c(1, 2, 3, 1, 5)
is_max(vec)
}
