% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_min}
\alias{is_min}
\title{is_min: Check if elements of an atomic vector are equal to its minimum value.}
\usage{
is_min(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector.}

\item{na.rm}{Logical; should NA values be removed before calculating the minimum value (default is TRUE).}
}
\value{
A logical vector with TRUE where \code{x} equals its minimum value.
}
\description{
This function accepts an atomic vector \code{x} and returns a logical vector with
TRUE where \code{x} equals its minimum value.
}
\examples{
vec = c(1, 2, 3, 1, 5)
is_min(vec)
}
