% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{Log-likelihood of a sample under the chi-squared distribution}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{A numeric vector of sample values.}

\item{df}{The degrees of freedom of the chi-squared distribution.}
}
\value{
The log-likelihood of the sample under the chi-squared distribution.
}
\description{
Log-likelihood of a sample under the chi-squared distribution
}
\examples{
x = rchisq(10, df = 5)
log_likelihood_chisq(x, df = 5)
}
