% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{Log-likelihood of a sample under the F distribution}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{A numeric vector of sample values.}

\item{df1}{The first degrees of freedom parameter of the F distribution.}

\item{df2}{The second degrees of freedom parameter of the F distribution.}
}
\value{
The log-likelihood of the sample under the F distribution.
}
\description{
Log-likelihood of a sample under the F distribution
}
\examples{
x = rf(10, df1 = 5, df2 = 10)
log_likelihood_f(x, df1 = 5, df2 = 10)
}
