% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Log-Likelihood of a sample under the normal distribution}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{A numeric vector of sample values.}

\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}
}
\value{
The log-likelihood of the sample under the normal distribution.
}
\description{
Log-Likelihood of a sample under the normal distribution
}
\examples{
x = rnorm(10, mean = 5, sd = 2)
log_likelihood_norm(x, mean = 5, sd = 2)
}
