% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Log-likelihood of a sample under the T distribution}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{A numeric vector of sample values.}

\item{df}{The degrees of freedom of the t distribution.}
}
\value{
The log-likelihood of the sample under the t distribution.
}
\description{
Log-likelihood of a sample under the T distribution
}
\examples{
x = rt(10, df = 5)
log_likelihood_t(x, df = 5)
}
