% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{Log-likelihood of a sample under the uniform distribution}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{A numeric vector of sample values.}

\item{min}{The minimum value of the uniform distribution.}

\item{max}{The maximum value of the uniform distribution.}
}
\value{
The log-likelihood of the sample under the uniform distribution.
}
\description{
Log-likelihood of a sample under the uniform distribution
}
\examples{
x = runif(10, min = 0, max = 1)
log_likelihood_unif(x, min = 0, max = 1)
}
