% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{r2}
\alias{r2}
\title{r2: Calculate the r-squared statistic between predicted and ground truth continuous variables.}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{A numeric vector of predicted values.}

\item{truth}{A numeric vector of ground truth values.}
}
\value{
The R-squared statistic between the predicted and ground truth values.
}
\description{
This function calculates the R-squared statistic (coefficient of determination)
between the predicted values and the ground truth continuous variables.
}
\examples{
pred = c(2, 4, 6, 8)
truth = c(2.2, 3.8, 6.1, 7.9)
r2(pred, truth)
}
