% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rando}
\alias{rando}
\title{Rando: A wrapper around sample for atomic vectors and data frame-like objects}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{An atomic vector or data frame-like object (data frame, matrix, or list).}

\item{n}{The number of samples or rows to return (default is 1).}

\item{replace}{Logical; should sampling be with replacement (default is TRUE).}
}
\value{
A vector or data frame-like object containing the sampled elements or rows.
}
\description{
This function checks whether the input \code{x} is an atomic vector or a
data frame-like object (data frame, matrix, or list) and returns either
\code{n} samples or \code{n} rows accordingly. If the input does not match any of these
types, an error message will be shown.
}
\examples{
vec = c(1, 2, 3, 4, 5)
df = data.frame(A = c(1, 2, 3), B = c(4, 5, 6))

# Test the rando function with an atomic vector and a data frame
rando(vec, n = 3, replace = T)
rando(df, n = 2, replace = T)
}
