% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{recall}
\alias{recall}
\title{Calculate the recall of a classifier}
\usage{
recall(pred, truth)
}
\arguments{
\item{pred}{A binary (logical) vector of predicted labels, where 1 (TRUE) represents a positive prediction and 0 (FALSE) represents a negative prediction.}

\item{truth}{A binary vector of true labels, where 1 represents a positive instance and 0 represents a negative instance.}
}
\value{
The recall of the classifier, as a numeric value between 0 and 1.
}
\description{
This function computes the recall (also known as sensitivity or true positive rate) of a classifier.
Recall is the proportion of true positive predictions among the actual positive instances (# True Positives / # All Positives).
}
\examples{
# Test the recall function
pred = c(1, 0, 1, 1, 0)
truth = c(1, 0, 1, 0, 1)
recall(pred, truth)

}
