% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{sensitivity}
\alias{sensitivity}
\title{Calculate the sensitivity of a classifier.}
\usage{
sensitivity(pred, truth)
}
\arguments{
\item{pred}{A binary (logical) vector of predicted labels, where 1 (TRUE) represents a positive prediction and 0 (FALSE) represents a negative prediction.}

\item{truth}{A binary (logical) vector of true labels, where 1 (TRUE) represents a positive instance and 0 (FALSE) represents a negative instance.}
}
\value{
The sensitivity of the classifier, as a numeric value between 0 and 1.
}
\description{
This function computes the sensitivity (also known as recall or true positive rate) of a classifier.
Sensitivity is the proportion of true positive predictions among the actual positive instances (# True Positives / # All Positives).
}
\examples{
# Test the sensitivity function
pred = c(1, 0, 1, 1, 0)
truth = c(1, 0, 1, 0, 1)
sensitivity(pred, truth)

}
