% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{specificity}
\alias{specificity}
\title{Calculate the specificity of a classifier.}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{A binary (logical) vector of predicted labels, where 1 (TRUE) represents a positive prediction and 0 (FALSE) represents a negative prediction.}

\item{truth}{A binary (logical) vector of true labels, where 1 (TRUE) represents a positive instance and 0 (FALSE) represents a negative instance.}
}
\value{
The specificity of the classifier, as a numeric value between 0 and 1.
}
\description{
This function computes the specificity (also known as true negative rate) of a classifier.
Specificity is the proportion of true negative predictions among the actual negative instances (# True Negatives / # All Negatives).
}
\examples{
# Test the specificity function
pred = c(1, 0, 1, 1, 0)
truth = c(1, 0, 1, 0, 1)
specificity(pred, truth)

}
